Move common test setup to BaseTableRouteTest
